refactor: move to types namespace
