+++ Arbiter.Net/Server/Messages/ServerShowMenuMessage.cs
@@ -1,5 +1,6 @@
 ﻿using Arbiter.Net.Annotations;
 using Arbiter.Net.Serialization;
+using Arbiter.Net.Server.Types;
 using Arbiter.Net.Types;
 
 namespace Arbiter.Net.Server.Messages;
