+++ Arbiter.Net/Server/Types/ServerEntityObject.cs
@@ -1,4 +1,4 @@
-﻿namespace Arbiter.Net.Server;
+﻿namespace Arbiter.Net.Server.Types;
 
 public class ServerEntityObject
 {
