+++ Arbiter.Net/Server/Types/ServerGroupBox.cs
@@ -1,4 +1,4 @@
-﻿namespace Arbiter.Net.Server;
+﻿namespace Arbiter.Net.Server.Types;
 
 public class ServerGroupBox
 {
